+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "firebase": "^11.0.0",
+    "js-cookie": "^3.0.5",
     "lucide-react": "^0.453.0",
     "next": "14.2.15",
     "react": "^18",
@@ -17,6 +18,7 @@
     "zustand": "^5.0.0"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.6",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
